requirements.txt
flask
flask-cors
python-dotenv
pyjwt
gradio
requests
